Se le cambió los nombres a ingles
